Role Management Updates

+++ MyServices/Impl/RoleManagement/RoleServiceImpl.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using AutoMapper;
+using HaloBiz.Data;
 using HaloBiz.DTOs.ApiDTOs;
 using HaloBiz.DTOs.ReceivingDTOs.RoleManagement;
 using HaloBiz.DTOs.TransferDTOs.RoleManagement;
@@ -19,19 +20,28 @@ namespace HaloBiz.MyServices.Impl.RoleManagement
         private readonly IModificationHistoryRepository _historyRepo;
         private readonly IMapper _mapper;
         private readonly IRoleRepository _roleRepo;
+        private readonly DataContext _context;
 
         public RoleServiceImpl(
             IModificationHistoryRepository historyRepo, 
             IRoleRepository roleRepo, 
+            DataContext dataContext,
             IMapper mapper)
         {
             this._mapper = mapper;
             this._historyRepo = historyRepo;
+            this._context = dataContext;
             this._roleRepo = roleRepo;
         }
 
         public async Task<ApiResponse> AddRole(HttpContext context, RoleReceivingDTO roleReceivingDTO)
         {
+            var item = await _roleRepo.FindRoleByName(roleReceivingDTO.Name);
+            if (item != null)
+            {
+                return new ApiResponse(400, "Role Name Already Exists.");
+            }
+
             var role = _mapper.Map<Role>(roleReceivingDTO);
             var savedRole = await _roleRepo.SaveRole(role);
             if (savedRole == null)
@@ -51,6 +61,17 @@ namespace HaloBiz.MyServices.Impl.RoleManagement
             }
             var roleTransferDTO = _mapper.Map<IEnumerable<RoleTransferDTO>>(roles);
             return new ApiOkResponse(roleTransferDTO);
+        }
+
+        public async Task<ApiResponse> GetAllClaims()
+        {
+            var claims = await _roleRepo.FindAllClaims();
+            if (claims == null)
+            {
+                return new ApiResponse(404);
+            }
+            var claimTransferDTO = _mapper.Map<IEnumerable<ClaimTransferDTO>>(claims);
+            return new ApiOkResponse(claimTransferDTO);
         }
 
         public async Task<ApiResponse> GetRoleById(long id)
@@ -82,30 +103,43 @@ namespace HaloBiz.MyServices.Impl.RoleManagement
             {
                 return new ApiResponse(404);
             }
-            
+
+            if(roleToUpdate.Name != roleReceivingDTO.Name)
+            {
+                var item = await _roleRepo.FindRoleByName(roleReceivingDTO.Name);
+                if (item != null)
+                {
+                    return new ApiResponse(400, "Role Name Already Exists.");
+                }
+            }          
+
             var summary = $"Initial details before change, \n {roleToUpdate.ToString()} \n" ;
 
             roleToUpdate.Name = roleReceivingDTO.Name;
             roleToUpdate.Description = roleReceivingDTO.Description;
 
-            var roleClaimsDeleted = await _roleRepo.DeleteRoleClaims(roleToUpdate);
-            if (!roleClaimsDeleted)
-            {
-                return new ApiResponse(500);
-            }
-
-            var roleClaimsToSave = _mapper.Map<ICollection<RoleClaim>>(roleReceivingDTO.RoleClaims);
-
-            roleToUpdate.RoleClaims = roleClaimsToSave;
-
             var updatedRole = await _roleRepo.UpdateRole(roleToUpdate);
-
-            summary += $"Details after change, \n {updatedRole.ToString()} \n";
 
             if (updatedRole == null)
             {
                 return new ApiResponse(500);
             }
+
+            var roleClaimsDeleted = await _roleRepo.DeleteRoleClaims(updatedRole);
+            if (!roleClaimsDeleted)
+            {
+                return new ApiResponse(500);
+            }
+                       
+            var roleClaimsToSave = _mapper.Map<ICollection<RoleClaim>>(roleReceivingDTO.RoleClaims);
+            foreach (var item in roleClaimsToSave)
+            {
+                item.RoleId = updatedRole.Id;
+            }
+            _context.RoleClaims.AddRange(roleClaimsToSave);
+
+            summary += $"Details after change, \n {updatedRole.ToString()} \n";
+    
             ModificationHistory history = new ModificationHistory(){
                 ModelChanged = "Role",
                 ChangeSummary = summary,
@@ -115,7 +149,7 @@ namespace HaloBiz.MyServices.Impl.RoleManagement
 
             await _historyRepo.SaveHistory(history);
 
-            var roleTransferDTO = _mapper.Map<RoleTransferDTO>(updatedRole);
+            var roleTransferDTO = _mapper.Map<RoleTransferDTO>(await _roleRepo.FindRoleById(id));
             return new ApiOkResponse(roleTransferDTO);
 
         }
@@ -126,6 +160,11 @@ namespace HaloBiz.MyServices.Impl.RoleManagement
             if (roleToDelete == null)
             {
                 return new ApiResponse(404);
+            }
+
+            if(roleToDelete.Name == ClaimConstants.SUPER_ADMIN || roleToDelete.Name == ClaimConstants.UN_ASSIGNED)
+            {
+                return new ApiResponse(400);
             }
 
             if (!await _roleRepo.DeleteRole(roleToDelete))
